+++ src/main/java/com/ocosur/ocosystem/dto/BatchUpdateDTO.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 import lombok.Data;
 @Data
 public class BatchUpdateDTO {
-    private Integer branchId;
+    private Long branchId;
     private BigDecimal kgTotal;
     private BigDecimal pricePerKg;
     private LocalDate date;
