+++ src/main/java/com/ocosur/ocosystem/repository/LossesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ocosur.ocosystem.model.Losses;
 
-public interface LossesRepository extends JpaRepository<Losses, Integer> {
+public interface LossesRepository extends JpaRepository<Losses, Long> {
     
 }
